+++ src/Controller/BlogController.php
@@ -3,19 +3,19 @@
 declare(strict_types=1);
 namespace App\Controller;
 
-use App\Model\Article;
+use App\Model\ArticleModel;
 use App\Services\TwigService;
 
 class BlogController
 {
 
     private TwigService $twig;
-    private Article $articles;
+    private ArticleModel $articles;
 
     public function __construct()
     {
         $this->twig = TwigService::getInstance();
-        $this->articles = new Article();
+        $this->articles = new ArticleModel();
     }
 
     public function showBlog()
