+++ src/View/article.html.twig
@@ -13,12 +13,11 @@
             </div>
             <!-- Portfolio Grid Items-->
             <div class="row justify-content-center">
-                    <div class="col-md-6 col-lg-4 mb-5">
-                        <div class="portfolio-item mx-auto">
+                    <div>
                             <h2>{{ article.title }}</h2><br>
                             <p><span>{{ article.header }}</span></p>
+                            <p>{{ article.content|nl2br   }}</p>
                             <h5>Modifié le {{ article.updated_at|date("d/m/Y") }}</h5>
-                        </div>
                     </div>
             </div>
         </div>
